Update package.json
There seems to be a redundancy about the baucis dependency declaration. I end up with two baucis subfolders in my node_modules, one explicit, one in baucis-swagger. Using only the peer dependency solves the problem.
Furthermore, there is a discrepancy about the dependency version. I updated it with the upper.